add insight server layer
