+++ core/src/main/scala/zio/insight/server/InsightServerConfig.scala
@@ -3,4 +3,5 @@ package zio.insight.server
 import java.time.Duration
 
 final case class InsightServerConfig(
+  port: Int,
   interval: Duration)
