+++ core/src/main/scala/zio/insight/server/package.scala
@@ -1,0 +1,10 @@
+package zio.insight
+
+import zio.ZLayer
+
+package object server {
+
+  def insightServerLayer(config: InsightServerConfig): ZLayer[Any, Throwable, Unit] =
+    ZLayer.fromZIO(InsightServer.make(config))
+
+}
