+++ core/src/test/scala/sample/SampleApp.scala
@@ -1,92 +1,15 @@
 package sample
 
 import zio._
-import zio.json._
-import zio.metrics.connectors.MetricsConfig
-import zio.metrics.connectors.insight.ClientMessage
-import zio.metrics.connectors.insight.ClientMessage.encAvailableMetrics
-import zio.metrics.connectors.insight.InsightPublisher
-import zio.metrics.jvm.DefaultJvmMetrics
-
-import zhttp.html._
-import zhttp.http._
-import zhttp.service.EventLoopGroup
-import zhttp.service.Server
-import zhttp.service.server.ServerChannelFactory
+import zio.insight.server.{insightServerLayer, InsightServerConfig}
 
 object SampleApp extends ZIOAppDefault with InstrumentedSample {
 
-  private val bindPort = 8080
-  private val nThreads = 5
+  private val insightServerConfig = InsightServerConfig(8080, 5.seconds)
 
-  private val metricsConfig = ZLayer.succeed(MetricsConfig(5.seconds))
-
-  private lazy val indexPage =
-    """<html>
-      |<title>Simple Server</title>
-      |<body>
-      |<p><a href="/metrics">Metrics</a></p>
-      |<p><a href="/insight/keys">Insight Metrics: Get all keys</a></p>
-      |</body
-      |</html>""".stripMargin
-
-  private lazy val static =
-    Http.collect[Request] { case Method.GET -> !! => Response.html(Html.fromString(indexPage)) }
-
-  private lazy val insightAllKeysRouter =
-    Http.collectZIO[Request] { case Method.GET -> !! / "insight" / "keys" =>
-      ZIO.serviceWithZIO[InsightPublisher](_.getAllKeys.map(_.toJson).map(data => noCors(Response.json(data))))
-    }
-
-  // POST: /insight/metrics body Seq[MetricKey] => Seq[MetricsNotification]
-  // TODO: Should we add an additional module with a layer implementation for zio-http?
-  // should be added (at some point) to zio-http ...
-  private lazy val insightGetMetricsRouter =
-    Http.collectZIO[Request] { case req @ Method.POST -> !! / "insight" / "metrics" =>
-      for {
-        request  <- req.body.asString.map(_.fromJson[ClientMessage.MetricsSelection])
-        response <- request match {
-                      case Left(e)  =>
-                        ZIO
-                          .debug(s"Failed to parse the input: $e")
-                          .as(
-                            Response.text(e).setStatus(Status.BadRequest),
-                          )
-                      case Right(r) =>
-                        ZIO
-                          .serviceWithZIO[InsightPublisher](_.getMetrics(r.selection))
-                          .map(_.toJson)
-                          .map(data => noCors(Response.json(data)))
-                    }
-      } yield response
-    }
-
-  private def noCors(r: Response): Response =
-    r.updateHeaders(_.combine(Headers(("Access-Control-Allow-Origin", "*"))))
-
-  private val server =
-    Server.port(bindPort) ++ Server.app(static ++ insightAllKeysRouter ++ insightGetMetricsRouter)
-
-  private lazy val runHttp = (server.start *> ZIO.never).forkDaemon
-
-  override def run: ZIO[Environment & ZIOAppArgs & Scope, Any, Any] = (for {
-    f <- runHttp
-    _ <- program
-    _ <- f.join
-  } yield ())
-    .provide(
-      ServerChannelFactory.auto,
-      EventLoopGroup.auto(nThreads),
-
-      // This is the general config for all backends
-      metricsConfig,
-
-      // The inisght reporting layer
-      zio.metrics.connectors.insight.insightLayer,
-
-      // Enable the ZIO internal metrics and the default JVM metricsConfig
-      // Do NOT forget the .unit for the JVM metrics layer
-      Runtime.enableRuntimeMetrics,
-      DefaultJvmMetrics.live.unit,
-    )
+  override def run: ZIO[Any with ZIOAppArgs with Scope, Any, Any] =
+    program
+      .provide(
+        insightServerLayer(insightServerConfig),
+      )
 }
