+++ project/Version.scala
@@ -7,5 +7,5 @@ object Version {
   val zio                  = "2.0.4"
   val zioJson              = "0.3.0"
   val zioHttp              = "2.0.0-RC11"
-  val zioMetricsConnectors = "2.0.3"
+  val zioMetricsConnectors = "2.0.4"
 }
